refactor: modularize Hacker News app into multiple files and folders
The original  file has been refactored and split into multiple smaller, more manageable files and folders. This modularization enhances maintainability, readability, and scalability of the application.

The new structure is as follows:

- : Configuration constants including URLs, store structure, and pagination settings.
- : Logic for fetching data from external APIs, including the  function.
- : Routing logic, including hash parsing and page routing functionality.
- : Helper functions responsible for updating the view with news feeds, news details, and comments.
- : Handles the rendering of the home page or the list of news feeds.
- : Responsible for rendering a specific page displaying news feeds with pagination.
- : Displays the detailed view of an individual news item.
- : Includes TypeScript interfaces to define the structure of data (, , , etc.).

This refactor achieves better separation of concerns and enhances the clarity and testability of the code. Each file is now responsible for a specific part of the application's functionality, making it easier to maintain, extend, and debug. Additionally, it simplifies the process of scaling the app with new features or changes.

Key Benefits:
- API functions and types have been moved into  and .
- View rendering logic is now contained within .
- Routing logic is more centralized in .
- News feed and detail views are now more isolated and modular, residing in .
- Application constants like URLs and the store structure have been moved to .

This modular approach ensures improved code organization and makes it easier to scale the app in the future.
