+++ src/page/index.ts
@@ -1,0 +1,2 @@
+export { default as newsDetail } from './news-detail-view'
+export { default as newsFeeds } from './news-feed-view'
